Modifies some functions to make them reusable and improves the situation of code copy pasting.
Adds 'click' listeners to all '.side-bar-btn' to change content of
'#content' when clicked.

Adds another 'click' listener to '#todays-new-task-bar' that loads
a form inside '#content' to collect information for a new today's task
and add it. It removes all the children of '#content' except
'.content-header-div' and replaces them with the form.

Modifies 'content-style.css' file to make some stylings more general.